+++ src/vips-emscripten.cpp
@@ -1105,6 +1105,7 @@ EMSCRIPTEN_BINDINGS(my_module) {
         .property("yres", &Image::yres)
         .property("filename", &Image::filename)
         // Handwritten setters
+        .property("kill", &Image::is_killed, &Image::set_kill)
         .property("onProgress", &Image::stub_getter,
                   &Image::set_progress_callback)
         // Auto-generated (class-)functions